requirements.txt
pymongo==3.10.1
fastapi
uvicorn
pyyaml
pytest
responses
factory-boy
pytest-mock
